+++ src/modules/admin/widgets/forms/PasswordRecoverActiveForm.php
@@ -5,24 +5,23 @@ namespace davidhirtz\yii2\skeleton\modules\admin\widgets\forms;
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\models\forms\PasswordRecoverForm;
 use davidhirtz\yii2\skeleton\modules\admin\widgets\forms\traits\EmailFieldTrait;
-use davidhirtz\yii2\skeleton\modules\admin\widgets\forms\traits\SubmitButtonTrait;
 use davidhirtz\yii2\skeleton\widgets\fontawesome\ActiveForm;
 use Yii;
 
-class PasswordRecoverFormActiveForm extends ActiveForm
+class PasswordRecoverActiveForm extends ActiveForm
 {
     use EmailFieldTrait;
-    use SubmitButtonTrait;
-
-    public ?PasswordRecoverForm $model = null;
 
     public $enableClientValidation = false;
+
+    public function __construct(public PasswordRecoverForm $model, $config = [])
+    {
+        parent::__construct($config);
+    }
 
     public function init(): void
     {
         $this->id = $this->getId(false) ?? 'password-recover-form';
-        $this->model ??= Yii::createObject(PasswordRecoverForm::class);
-
         parent::init();
     }
 
@@ -43,10 +42,5 @@ class PasswordRecoverFormActiveForm extends ActiveForm
     {
         $content = Yii::t('skeleton', 'Enter your email address and we will send you instructions how to reset your password.');
         return Html::tag('p', $content);
-    }
-
-    public function sendEmailButton(): string
-    {
-        return $this->submitButton(Yii::t('skeleton', 'Send Email'));
     }
 }