+++ src/modules/admin/widgets/forms/traits/EmailFieldTrait.php
@@ -2,11 +2,14 @@
 
 namespace davidhirtz\yii2\skeleton\modules\admin\widgets\forms\traits;
 
+use Yii;
 use yii\widgets\ActiveField;
 
 trait EmailFieldTrait
 {
-    public function emailField(): ActiveField|string
+    use SubmitButtonTrait;
+
+    public function emailField(array $options = []): ActiveField|string
     {
         $field = $this->field($this->model, 'email', [
             'icon' => 'envelope',
@@ -14,9 +17,15 @@ trait EmailFieldTrait
         ]);
 
         return $field->textInput([
+            ...$options,
             'autocomplete' => 'username',
             'autofocus' => !$this->model->hasErrors(),
             'type' => 'email',
         ]);
+    }
+
+    public function sendEmailButton(): string
+    {
+        return $this->submitButton(Yii::t('skeleton', 'Send Email'));
     }
 }